requirement.txt
Django==4.2.11
pytest-django==4.8.0